refactor : user, auth 도메인 생성 & 도메인 구조로 변경
