+++ src/main/java/space/space_spring/domain/authorization/jwt/model/TokenType.java
@@ -1,0 +1,6 @@
+package space.space_spring.domain.authorization.jwt.model;
+
+public enum TokenType {
+    REFRESH,
+    ACCESS
+}
