--- src/main/java/space/space_spring/dto/jwt/TokenType.java
@@ -1,6 +1,0 @@
-package space.space_spring.dto.jwt;
-
-public enum TokenType {
-    REFRESH,
-    ACCESS
-}
